common fixture app_env
