+++ services/web/server/tests/unit/with_dbs/03/invitations/test_login_registration_invitations.py
@@ -11,40 +11,11 @@ from models_library.api_schemas_invitations.invitations import ApiInvitationCont
 from pytest_mock import MockerFixture
 from pytest_simcore.aioresponses_mocker import AioResponsesMock
 from pytest_simcore.helpers.utils_assert import assert_status
-from pytest_simcore.helpers.utils_envs import EnvVarsDict, setenvs_from_dict
 from simcore_service_webserver.login.handlers_registration import (
     InvitationCheck,
     InvitationInfo,
 )
 from simcore_service_webserver.login.settings import LoginSettingsForProduct
-
-
-@pytest.fixture
-def app_environment(
-    app_environment: EnvVarsDict,
-    env_devel_dict: EnvVarsDict,
-    monkeypatch: pytest.MonkeyPatch,
-) -> EnvVarsDict:
-    login_envs = setenvs_from_dict(
-        monkeypatch,
-        {
-            "LOGIN_REGISTRATION_CONFIRMATION_REQUIRED": "1",
-            "LOGIN_REGISTRATION_INVITATION_REQUIRED": "0",
-            "LOGIN_2FA_CODE_EXPIRATION_SEC": "60",
-        },
-    )
-
-    # set INVITATIONS_* variables using those in .devel-env
-    invitation_envs = setenvs_from_dict(
-        monkeypatch,
-        envs={
-            name: value
-            for name, value in env_devel_dict.items()
-            if name.startswith("INVITATIONS_")
-        },
-    )
-
-    return app_environment | login_envs | invitation_envs
 
 
 async def test_check_registration_invitation_when_not_required(
@@ -69,7 +40,7 @@ async def test_check_registration_invitation_when_not_required(
     data, _ = await assert_status(response, web.HTTPOk)
 
     invitation = InvitationInfo.parse_obj(data)
-    assert invitation.email == None
+    assert invitation.email is None
 
 
 async def test_check_registration_invitations_with_old_code(
@@ -80,7 +51,7 @@ async def test_check_registration_invitations_with_old_code(
     mocker.patch(
         "simcore_service_webserver.login.handlers_registration.get_plugin_settings",
         autospec=True,
-        return_value=LoginSettingsForProduct(
+        return_value=LoginSettingsForProduct.create_from_envs(
             LOGIN_REGISTRATION_INVITATION_REQUIRED=True,  # <--
         ),
     )
@@ -92,10 +63,10 @@ async def test_check_registration_invitations_with_old_code(
     data, _ = await assert_status(response, web.HTTPOk)
 
     invitation = InvitationInfo.parse_obj(data)
-    assert invitation.email == None
+    assert invitation.email is None
 
 
-@pytest.mark.acceptance_test
+@pytest.mark.acceptance_test()
 async def test_check_registration_invitation_and_get_email(
     client: TestClient,
     mocker: MockerFixture,
@@ -106,7 +77,7 @@ async def test_check_registration_invitation_and_get_email(
     mocker.patch(
         "simcore_service_webserver.login.handlers_registration.get_plugin_settings",
         autospec=True,
-        return_value=LoginSettingsForProduct(
+        return_value=LoginSettingsForProduct.create_from_envs(
             LOGIN_REGISTRATION_INVITATION_REQUIRED=True,  # <--
         ),
     )
